autocomplete for real, part 1
