+++ src/agents/tools/completers/mod.rs
@@ -1,0 +1,14 @@
+//! Standardized option completers, which are used by a lot of commands.
+//! 
+//! Demonstrationg completions is also pretty hard, because where the cursor is changes things.
+//! The format used in this module's documentation is:
+//! 
+//! - `command text|` -> explanation of what happens
+//! 
+//! where `|` is the cursor's location when you press tab.
+//! There'll also be an explanation of the command, or it'll be a common \*nix one.
+
+mod gnu;
+// pub use gnu::GnuCompleter;
+mod bsd;
+pub use bsd::BsdCompleter;
